Unify ring logic and make it testable; clarify blocking/signaling.
