+++ src/rtl_sdr_fm.cpp
@@ -34,6 +34,8 @@
 #include <unistd.h>
 #include "dsd.h"
 #include "dsp/simd_widen.h"
+#include "runtime/input_ring.h"
+#include "runtime/ring.h"
 
 /*
  * Environment variables (runtime configuration)
@@ -787,218 +789,14 @@ demod_mt_run_two(struct demod_state* s, void (*f0)(void*), void* a0, void (*f1)(
     pthread_mutex_unlock(&s->mt_lock);
 }
 
-struct output_state {
-    int rate;
-    int16_t* buffer;
-    size_t capacity;
-    std::atomic<size_t> head;
-    std::atomic<size_t> tail;
-    pthread_cond_t ready;
-    pthread_cond_t space;
-    pthread_mutex_t ready_m;
-};
 
-/* Simple SPSC ring for interleaved I/Q int16_t samples (input path) */
-struct input_ring_state {
-    int16_t* buffer;
-    size_t capacity; /* in int16_t elements */
-    std::atomic<size_t> head;
-    std::atomic<size_t> tail;
-    pthread_cond_t ready;
-    pthread_mutex_t ready_m;
-};
 
-/**
- * Number of samples currently in the input ring.
- */
-static inline size_t
-input_ring_used(const struct input_ring_state* r) {
-    size_t h = r->head.load();
-    size_t t = r->tail.load();
-    if (h >= t) {
-        return h - t;
-    }
-    return r->capacity - (t - h);
-}
 
-/**
- * Number of free slots available for writing in the input ring.
- */
-static inline size_t
-input_ring_free(const struct input_ring_state* r) {
-    return (r->capacity - 1) - input_ring_used(r);
-}
 
-/**
- * Check if the input ring is empty.
- */
-static inline int
-input_ring_is_empty(const struct input_ring_state* r) {
-    return r->head.load() == r->tail.load();
-}
 
-/**
- * Clear the input ring head/tail indices.
- */
-static inline void
-input_ring_clear(struct input_ring_state* r) {
-    r->tail.store(0);
-    r->head.store(0);
-}
 
-/**
- * Reserve up to two contiguous writable regions totaling at least min_needed
- * (or less if near full). When the ring is full, this function grants zero
- * space to avoid advancing the consumer. The producer may choose to drop or
- * skip the current block.
- *
- * @param r          Input ring buffer state.
- * @param min_needed Minimum samples requested for writing.
- * @param p1         [out] First writable region pointer or NULL.
- * @param n1         [out] First writable region length.
- * @param p2         [out] Second writable region pointer or NULL.
- * @param n2         [out] Second writable region length.
- * @return Total writable samples granted across regions.
- */
-static int
-input_ring_reserve(struct input_ring_state* r, size_t min_needed, int16_t** p1, size_t* n1, int16_t** p2, size_t* n2) {
-    size_t free_sp = input_ring_free(r);
-    /* Producer must never advance consumer tail; if full, grant nothing */
-    /* Provide up to min(free_sp, min_needed) across at most two regions */
-    size_t grant = (min_needed < free_sp) ? min_needed : free_sp;
-    size_t h = r->head.load();
-    size_t first = r->capacity - h;
-    if (first > grant) {
-        first = grant;
-    }
-    *p1 = (first > 0) ? (r->buffer + h) : NULL;
-    *n1 = first;
-    *p2 = NULL;
-    *n2 = 0;
-    if (grant > first) {
-        *p2 = r->buffer;
-        *n2 = grant - first;
-    }
-    return (int)(*n1 + *n2);
-}
 
-/**
- * Commit produced samples and signal consumer on empty→non-empty transition.
- *
- * @param r         Input ring buffer state.
- * @param produced  Number of samples produced to commit.
- */
-static void
-input_ring_commit(struct input_ring_state* r, size_t produced) {
-    if (produced == 0) {
-        return;
-    }
-    int need_signal = input_ring_is_empty(r);
-    size_t h = r->head.load();
-    h += produced;
-    if (h >= r->capacity) {
-        h %= r->capacity;
-    }
-    r->head.store(h);
-    if (need_signal) {
-        pthread_mutex_lock(&r->ready_m);
-        pthread_cond_signal(&r->ready);
-        pthread_mutex_unlock(&r->ready_m);
-    }
-}
 
-/**
- * Write up to count samples into the input ring buffer, dropping oldest half
- * when necessary to avoid blocking.
- *
- * @param r     Input ring buffer state.
- * @param data  Source samples to write.
- * @param count Number of samples to write.
- */
-static void
-input_ring_write(struct input_ring_state* r, const int16_t* data, size_t count) {
-    int need_signal = input_ring_is_empty(r);
-    while (count > 0 && !exitflag) {
-        size_t free_sp = input_ring_free(r);
-        if (free_sp == 0) {
-            /* Ring full: to avoid racing the consumer, drop remainder */
-            break;
-        }
-        size_t write_now = (count < free_sp) ? count : free_sp;
-        size_t h = r->head.load();
-        size_t first = r->capacity - h;
-        if (first > write_now) {
-            first = write_now;
-        }
-        memcpy(r->buffer + h, data, first * sizeof(int16_t));
-        if (write_now > first) {
-            memcpy(r->buffer, data + first, (write_now - first) * sizeof(int16_t));
-            h = write_now - first;
-        } else {
-            h += first;
-            if (h == r->capacity) {
-                h = 0;
-            }
-        }
-        r->head.store(h);
-        data += write_now;
-        count -= write_now;
-    }
-    if (need_signal) {
-        pthread_mutex_lock(&r->ready_m);
-        pthread_cond_signal(&r->ready);
-        pthread_mutex_unlock(&r->ready_m);
-    }
-}
-
-/**
- * Read up to max_count samples from input ring buffer, blocking with timeout
- * until data is available or exit is requested.
- *
- * @param r         Input ring buffer state.
- * @param out       Destination buffer for samples.
- * @param max_count Maximum number of samples to read.
- * @return Number of samples read (>=1), 0 if max_count is 0, or -1 on exit.
- */
-static int
-input_ring_read_block(struct input_ring_state* r, int16_t* out, size_t max_count) {
-    if (max_count == 0) {
-        return 0;
-    }
-    while (input_ring_is_empty(r)) {
-        struct timespec ts;
-        clock_gettime(CLOCK_REALTIME, &ts);
-        ts.tv_nsec += 10L * 1000000L; /* 10ms */
-        if (ts.tv_nsec >= 1000000000L) {
-            ts.tv_sec += ts.tv_nsec / 1000000000L;
-            ts.tv_nsec = ts.tv_nsec % 1000000000L;
-        }
-        pthread_mutex_lock(&r->ready_m);
-        pthread_cond_timedwait(&r->ready, &r->ready_m, &ts);
-        pthread_mutex_unlock(&r->ready_m);
-        if (exitflag) {
-            return -1;
-        }
-    }
-    size_t available = input_ring_used(r);
-    size_t read_now = (max_count < available) ? max_count : available;
-    size_t t = r->tail.load();
-    size_t first = r->capacity - t;
-    if (first > read_now) {
-        first = read_now;
-    }
-    memcpy(out, r->buffer + t, first * sizeof(int16_t));
-    t += first;
-    if (t == r->capacity) {
-        t = 0;
-    }
-    if (read_now > first) {
-        memcpy(out + first, r->buffer, (read_now - first) * sizeof(int16_t));
-        t = read_now - first;
-    }
-    r->tail.store(t);
-    return (int)read_now;
-}
 
 struct controller_state {
     int exit_flag;
@@ -1018,14 +816,6 @@ struct output_state output;
 struct controller_state controller;
 static struct input_ring_state input_ring;
 
-#define safe_cond_signal(n, m)                                                                                         \
-    pthread_mutex_lock(m);                                                                                             \
-    pthread_cond_signal(n);                                                                                            \
-    pthread_mutex_unlock(m)
-#define safe_cond_wait(n, m)                                                                                           \
-    pthread_mutex_lock(m);                                                                                             \
-    pthread_cond_wait(n, m);                                                                                           \
-    pthread_mutex_unlock(m)
 
 /**
  * Optionally enable realtime scheduling and set CPU affinity for the current
@@ -1106,257 +896,14 @@ maybe_set_thread_realtime_and_affinity(const char* role) {
     }
 }
 
-/**
- * Number of samples currently in the output ring.
- *
- * @param o Output ring state.
- * @return Number of queued samples.
- */
-static inline size_t
-ring_used(const struct output_state* o) {
-    /* Atomics policy: head/tail are atomics. We use default sequential
-       consistency for simplicity. In an SPSC ring, this could be relaxed
-       to acquire/release without changing behavior. */
-    size_t h = o->head.load();
-    size_t t = o->tail.load();
-    if (h >= t) {
-        return h - t;
-    }
-    return o->capacity - (t - h);
-}
 
-/**
- * Number of free slots available in the output ring.
- *
- * @param o Output ring state.
- * @return Number of writable samples before the ring becomes full.
- */
-static inline size_t
-ring_free(const struct output_state* o) {
-    return (o->capacity - 1) - ring_used(o);
-}
 
-/**
- * Check if the output ring is empty.
- *
- * @param o Output ring state.
- * @return Non-zero if empty, zero otherwise.
- */
-static inline int
-ring_is_empty(const struct output_state* o) {
-    /* See atomics note in ring_used() for ordering considerations. */
-    return o->head.load() == o->tail.load();
-}
 
-/**
- * Clear the output ring head/tail indices.
- *
- * @param o Output ring state to clear.
- */
-static inline void
-ring_clear(struct output_state* o) {
-    /* Clearing indices; with relaxed ordering this would be a release store. */
-    o->tail.store(0);
-    o->head.store(0);
-}
 
-/* Write up to count samples, blocking until space is available. Signals data availability after writes. */
-static void
-ring_write(struct output_state* o, const int16_t* data, size_t count) {
-    int need_signal = ring_is_empty(o);
-    while (count > 0 && !exitflag) {
-        size_t free_sp = ring_free(o);
-        if (free_sp == 0) {
-            /* Wait for space with timeout to avoid indefinite blocking */
-            struct timespec ts;
-            clock_gettime(CLOCK_REALTIME, &ts);
-            ts.tv_nsec += 50L * 1000000L; /* 50ms */
-            if (ts.tv_nsec >= 1000000000L) {
-                ts.tv_sec += ts.tv_nsec / 1000000000L;
-                ts.tv_nsec = ts.tv_nsec % 1000000000L;
-            }
-            pthread_mutex_lock(&o->ready_m);
-            pthread_cond_timedwait(&o->space, &o->ready_m, &ts);
-            pthread_mutex_unlock(&o->ready_m);
-            if (exitflag) {
-                break;
-            }
-            continue;
-        }
-        size_t write_now = (count < free_sp) ? count : free_sp;
-        size_t h = o->head.load();
-        size_t first = o->capacity - h;
-        if (first > write_now) {
-            first = write_now;
-        }
-        memcpy(o->buffer + h, data, first * sizeof(int16_t));
-        if (write_now > first) {
-            memcpy(o->buffer, data + first, (write_now - first) * sizeof(int16_t));
-            h = write_now - first;
-        } else {
-            h += first;
-            if (h == o->capacity) {
-                h = 0;
-            }
-        }
-        o->head.store(h);
-        data += write_now;
-        count -= write_now;
-    }
-    if (need_signal) {
-        safe_cond_signal(&o->ready, &o->ready_m);
-    }
-}
 
-/* Same as ring_write but does not signal; caller decides when to signal */
-static void
-ring_write_no_signal(struct output_state* o, const int16_t* data, size_t count) {
-    while (count > 0 && !exitflag) {
-        size_t free_sp = ring_free(o);
-        if (free_sp == 0) {
-            /* Wait for space with timeout to avoid indefinite blocking */
-            struct timespec ts;
-            clock_gettime(CLOCK_REALTIME, &ts);
-            ts.tv_nsec += 50L * 1000000L; /* 50ms */
-            if (ts.tv_nsec >= 1000000000L) {
-                ts.tv_sec += ts.tv_nsec / 1000000000L;
-                ts.tv_nsec = ts.tv_nsec % 1000000000L;
-            }
-            pthread_mutex_lock(&o->ready_m);
-            pthread_cond_timedwait(&o->space, &o->ready_m, &ts);
-            pthread_mutex_unlock(&o->ready_m);
-            if (exitflag) {
-                break;
-            }
-            continue;
-        }
-        size_t write_now = (count < free_sp) ? count : free_sp;
-        size_t h = o->head.load();
-        size_t first = o->capacity - h;
-        if (first > write_now) {
-            first = write_now;
-        }
-        memcpy(o->buffer + h, data, first * sizeof(int16_t));
-        if (write_now > first) {
-            memcpy(o->buffer, data + first, (write_now - first) * sizeof(int16_t));
-            h = write_now - first;
-        } else {
-            h += first;
-            if (h == o->capacity) {
-                h = 0;
-            }
-        }
-        o->head.store(h);
-        data += write_now;
-        count -= write_now;
-    }
-}
 
-/**
- * Write and signal consumer only if the ring transitions from empty to
- * non-empty. Reduces unnecessary wakeups while ensuring timely reads.
- *
- * @param o     Output ring buffer state.
- * @param data  Source samples to write.
- * @param count Number of samples to write.
- */
-static void
-ring_write_signal_on_empty_transition(struct output_state* o, const int16_t* data, size_t count) {
-    int need_signal = ring_is_empty(o);
-    ring_write_no_signal(o, data, count);
-    if (need_signal) {
-        safe_cond_signal(&o->ready, &o->ready_m);
-    }
-}
 
-/**
- * Read one sample from the output ring, blocking with timeout until available
- * or exit requested.
- *
- * @param o    Output ring buffer state.
- * @param out  Destination for one sample.
- * @return 0 on success, -1 on exit.
- */
-static int
-ring_read_one(struct output_state* o, int16_t* out) {
-    while (ring_is_empty(o)) {
-        struct timespec ts;
-        clock_gettime(CLOCK_REALTIME, &ts);
-        ts.tv_nsec += 10L * 1000000L; /* 10ms */
-        if (ts.tv_nsec >= 1000000000L) {
-            ts.tv_sec += ts.tv_nsec / 1000000000L;
-            ts.tv_nsec = ts.tv_nsec % 1000000000L;
-        }
-        pthread_mutex_lock(&o->ready_m);
-        pthread_cond_timedwait(&o->ready, &o->ready_m, &ts);
-        pthread_mutex_unlock(&o->ready_m);
-        if (exitflag) {
-            return -1;
-        }
-    }
-    size_t t = o->tail.load();
-    *out = o->buffer[t];
-    t += 1;
-    if (t == o->capacity) {
-        t = 0;
-    }
-    o->tail.store(t);
-    /* Signal space available for producer */
-    safe_cond_signal(&o->space, &o->ready_m);
-    return 0;
-}
 
-/**
- * Read up to max_count samples into out. Blocks until at least one sample is available or exit.
- * Signals producer space once after the batch.
- *
- * @param o         Output ring buffer state.
- * @param out       Destination buffer for samples.
- * @param max_count Maximum number of samples to read.
- * @return Number of samples read (>=1) or -1 on exit.
- */
-static int
-ring_read_batch(struct output_state* o, int16_t* out, size_t max_count) {
-    if (max_count == 0) {
-        return 0;
-    }
-    while (ring_is_empty(o)) {
-        struct timespec ts;
-        clock_gettime(CLOCK_REALTIME, &ts);
-        ts.tv_nsec += 10L * 1000000L; /* 10ms */
-        if (ts.tv_nsec >= 1000000000L) {
-            ts.tv_sec += ts.tv_nsec / 1000000000L;
-            ts.tv_nsec = ts.tv_nsec % 1000000000L;
-        }
-        pthread_mutex_lock(&o->ready_m);
-        pthread_cond_timedwait(&o->ready, &o->ready_m, &ts);
-        pthread_mutex_unlock(&o->ready_m);
-        if (exitflag) {
-            return -1;
-        }
-    }
-
-    size_t available = ring_used(o);
-    size_t read_now = (max_count < available) ? max_count : available;
-    size_t t = o->tail.load();
-    size_t first = o->capacity - t;
-    if (first > read_now) {
-        first = read_now;
-    }
-    memcpy(out, o->buffer + t, first * sizeof(int16_t));
-    t += first;
-    if (t == o->capacity) {
-        t = 0;
-    }
-    if (read_now > first) {
-        memcpy(out + first, o->buffer, (read_now - first) * sizeof(int16_t));
-        t = read_now - first;
-    }
-    o->tail.store(t);
-    /* Signal space available once for the whole batch */
-    safe_cond_signal(&o->space, &o->ready_m);
-    return (int)read_now;
-}
 
 /* {length, coef, coef, coef}  and scaled by 2^15
    for now, only length 9, optimal way to get +85% bandwidth */
